C2SoftHevcDec: Initialize mStride after first header decode
After first header decode, decoder stride was set to ALIGN128(width),
but mStride that tracks the stride in the plugin wasn't set to this.

When allocated buffer has a stride that is different from
ALIGN128(width), this would lead to wrong stride used in the decoder.

Bug: 238984399
Test: atest CtsMediaV2TestCases -- --module-arg \
CtsMediaV2TestCases:instrumentation-arg:\
codec-prefix:=c2.android.hevc.decoder

Change-Id: I06886cc718a1b8b9fa247b248b23adb2feb80f90